+++ src/package.json
@@ -52,6 +52,8 @@
     "@graphql-codegen/typescript-urql-graphcache": "3.1.0",
     "@graphql-codegen/urql-introspection": "3.0.0",
     "@graphql-typed-document-node/core": "3.2.0",
+    "@headlessui/vue": "1.7.17",
+    "@heroicons/vue": "2.1.1",
     "@intlify/eslint-plugin-vue-i18n": "3.0.0-next.5",
     "@nuxtjs/eslint-config-typescript": "12.1.0",
     "@nuxtjs/i18n": "8.0.0",
