+++ src/pages/contact/index.vue
@@ -1,0 +1,3 @@
+<template>
+  <div />
+</template>
