+++ src/pages/faq/index.vue
@@ -45,8 +45,6 @@ import { marked } from 'marked'
 
 import type { CrealFaq } from '~/types/creal'
 
-definePageMeta({ colorMode: 'dark' })
-
 const { t, locale } = useI18n()
 const route = useRoute()
 const strapiFetch = useStrapiFetch()
