+++ src/pages/legal-notice.vue
@@ -1,9 +1,3 @@
 <template>
   <VioPageLegalNotice />
 </template>
-
-<script setup lang="ts">
-definePageMeta({
-  colorMode: 'dark',
-})
-</script>
