+++ src/pages/privacy-policy.vue
@@ -55,9 +55,3 @@
     }"
   />
 </template>
-
-<script setup lang="ts">
-definePageMeta({
-  colorMode: 'dark',
-})
-</script>
